Remove bda suffix and delete out-of-date files
